+++ stroom-core-client-widget/src/main/java/stroom/widget/popup/client/presenter/PopupView.java
@@ -40,6 +40,6 @@ public interface PopupView extends View, HasUiHandlers<PopupUiHandlers> {
     void hide(boolean autoClose);
 
     enum PopupType {
-        POPUP, DIALOG, CLOSE_DIALOG, OK_CANCEL_DIALOG
+        POPUP, DIALOG, CLOSE_DIALOG, OK_CANCEL_DIALOG, ACCEPT_REJECT_DIALOG
     }
 }