+++ stroom-core-server/src/main/java/stroom/node/server/ClientPropertiesServiceImpl.java
@@ -47,6 +47,15 @@ public class ClientPropertiesServiceImpl implements ClientPropertiesService {
         addProperty(props, ClientProperties.QUERY_INFO_POPUP_ENABLED);
         addProperty(props, ClientProperties.QUERY_INFO_POPUP_TITLE);
         addProperty(props, ClientProperties.QUERY_INFO_POPUP_VALIDATION_REGEX);
+        addProperty(props, ClientProperties.ACTIVITY_ENABLED);
+        addProperty(props, ClientProperties.ACTIVITY_CHOOSE_ON_STARTUP);
+        addProperty(props, ClientProperties.ACTIVITY_MANAGER_TITLE);
+        addProperty(props, ClientProperties.ACTIVITY_EDITOR_TITLE);
+        addProperty(props, ClientProperties.ACTIVITY_EDITOR_BODY);
+        addProperty(props, ClientProperties.SPLASH_ENABLED);
+        addProperty(props, ClientProperties.SPLASH_TITLE);
+        addProperty(props, ClientProperties.SPLASH_BODY);
+        addProperty(props, ClientProperties.SPLASH_VERSION);
         addProperty(props, ClientProperties.AUTHENTICATION_SERVICE_URL);
         addProperty(props, ClientProperties.ADVERTISED_HOST_URL);
         addProperty(props, ClientProperties.USERS_UI_URL);