+++ stroom-pipeline/src/main/java/stroom/pipeline/server/writer/HDFSFileAppender.java
@@ -31,6 +31,7 @@ import stroom.pipeline.server.factory.PipelineProperty;
 import stroom.pipeline.shared.ElementIcons;
 import stroom.pipeline.shared.data.PipelineElementType;
 import stroom.pipeline.shared.data.PipelineElementType.Category;
+import stroom.util.io.ByteCountOutputStream;
 import stroom.util.spring.StroomScope;
 
 import javax.inject.Inject;
@@ -71,6 +72,7 @@ public class HDFSFileAppender extends AbstractAppender {
 
     private Configuration conf;
     private UserGroupInformation userGroupInformation;
+    private ByteCountOutputStream byteCountOutputStream;
 
     @Inject
     HDFSFileAppender(final ErrorReceiverProxy errorReceiverProxy,
@@ -153,13 +155,18 @@ public class HDFSFileAppender extends AbstractAppender {
 
     @Override
     protected OutputStream createOutputStream() throws IOException {
+        byteCountOutputStream = new ByteCountOutputStream(createHDFSLockedOutputStream());
+        return byteCountOutputStream;
+    }
+
+    HDFSLockedOutputStream createHDFSLockedOutputStream() throws IOException {
         try {
             if (outputPaths == null || outputPaths.length == 0) {
                 throw new IOException("No output paths have been set");
             }
 
             // Get a path to use.
-            String path = null;
+            String path;
             if (outputPaths.length == 1) {
                 path = outputPaths[0];
             } else {
@@ -180,6 +187,14 @@ public class HDFSFileAppender extends AbstractAppender {
         } catch (final Exception e) {
             throw new IOException(e.getMessage(), e);
         }
+    }
+
+    @Override
+    long getCurrentOutputSize() {
+        if (byteCountOutputStream != null) {
+            return byteCountOutputStream.getCount();
+        }
+        return 0;
     }
 
     /**
@@ -256,7 +271,7 @@ public class HDFSFileAppender extends AbstractAppender {
 
         if (hdfsLockedOutputStream == null) {
             throw new RuntimeException(String.format(
-                    "Something went wrong creating the HDFSLockedOutputStream, lockFile %s, outFile %s, hdfs uri %s",
+                    "Something went wrong creating the HDFSLockedOutputStream, outFile %s, hdfs uri %s",
                     filePath, hdfsUri));
         }